Library cobol_typeck

Superbol is a set of utilities for COBOL developers, that can be selected by sub-commands.

This package contains data and type checking functions on COBOL data items.

The entry point of this library is the module: Cobol_typeck.